pretrain helper
